Remove hashtag triggers
Remove triggers related to 4edit summary hashtags. The database that was
once used for this has been inactive since 2018
(<https://phabricator.wikimedia.org/T188205>). There is a replacement
database, but it is not readily accessible outside of the hashtags Cloud
VPS project.

Reviving these triggers in the future is possible, but should be done in
cooperation with the maintainers of the hashtags Cloud VPS project,
possibly by exposing a new API on <https://hashtags.wmcloud.org/>.